// studio/datart-versions/datart-1.0.0-rc.1/frontend/src/app/types/ChartConfig.ts
export type FilterFieldAction = {
  condition?: FilterCondition;
  visibility?: FilterVisibility;
  facade?: FilterFacade;
  width?: string;
};

export type FilterVisibility =
  | Lowercase<ControllerVisibilityTypes>
  | {
      visibility: ControllerVisibilityTypes.Condition;
      fieldUid: string;
      relation: string;
      value: string;
    };

export type FilterFacade =
  | Uncapitalize<keyof typeof ControllerFacadeTypes>
  | {
      facade: Uncapitalize<keyof typeof ControllerFacadeTypes>;
      [key: string]: string | number;
    };

export type FilterCondition = {
  name: string;
  type: FilterConditionType;
  value?:
    | Lowercase<keyof typeof FilterRelationType>
    | string
    | number
    | [number, number]
    | string[]
    | Array<RelationFilterValue>
    | TimeFilterConditionValue;
  visualType: string;
  operator?:
    | string
    | Lowercase<keyof typeof FilterRelationType>
    | Uncapitalize<keyof typeof FilterSqlOperator>;
  children?: FilterCondition[];
};
// ...
export type ChartDataSectionField = {
  uid?: string;
  colName: string;
  desc?: string;
  type: DataViewFieldType;
  category: Uncapitalize<keyof typeof ChartDataViewFieldCategory>;
  expression?: string;
  field?: string;
  sort?: SortFieldAction;
  alias?: AliasFieldAction;
  format?: FormatFieldAction;
  aggregate?: AggregateFieldActionType;
  filter?: FilterFieldAction;
  color?: ColorFieldAction;
  size?: number;
  path?: string[];
};

export type SortFieldAction = {
  type: SortActionType;
  value?: any;
};

export type ColorFieldAction = {
  start?: string;
  end?: string;
  colors?: Array<{ key: string; value: string }>;
};

export type FormatFieldAction = {
  type: FieldFormatType;
  [FieldFormatType.Numeric]?: {
    decimalPlaces: number;
    unitKey?: NumberUnitKey;
    useThousandSeparator?: boolean;
    prefix?: string;
    suffix?: string;
  };
  [FieldFormatType.Currency]?: {
    decimalPlaces: number;
    unitKey?: NumberUnitKey;
    useThousandSeparator?: boolean;
    currency?: string;
  };
  [FieldFormatType.Percentage]?: {
    decimalPlaces: number;
  };
  [FieldFormatType.Scientific]?: {
    decimalPlaces: number;
  };
  [FieldFormatType.Date]?: {
    format: string;
  };
  [FieldFormatType.Custom]?: {
    format: string;
  };
};

export type AliasFieldAction = {
  name?: string;
  desc?: string;
};